Theming library implementation
Basically a copy-pasta from XesignPhotos at this point.
There's support for Dark and Light themes right now, but once the app has risen in terms of functionality to a basic close-to level of VW9, I might consider adding a Black theme.
We'll still need to expand this implementation to support Accent colors.