WebcamViewerX.ViewManagement project
This project will be used to switch Views in the main application.
Its implementation right now is pretty clean, and the main app will also try and use this in a simple way.
WV9 would have the Pages, the Frames and the transitions hard-coded into the main app. That was a really bad implementation.