WebcamViewerX.Settings project
This will hold the Settings view of course. This is also a demo/placeholder for now..
The configuration changing code will be in WebcamViewerX.Configuration, if needed.

Also note that the Webcam stuff, such as loading the image, saving it, caching etc... are going to be handled in a different class.